# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: M360 API
  version: 1.0.0
  description: |
    M360 SME Lending CRM API for Shariah-compliant Murabaha financing in Saudi Arabia.
    Bilingual Arabic/English with full loan lifecycle management, Saudi government
    integrations (SIMAH, Bayan, Nafath, Yaqeen, Watheq), and role-based access control.

    ## Roles
    - `admin` - Full system access
    - `manager` - Management-level access
    - `loan_officer` - Loan origination and management
    - `credit_analyst` - Credit assessment and committee packages
    - `compliance_officer` - Compliance oversight
    - `collections_officer` - Collections management
    - `data_entry` - Data entry operations

    ## Application Workflow
    Draft -> Submitted -> PreApproved -> DocsCollected -> CreditAssessment -> CommitteeReview -> Approved/Rejected -> Disbursed
  contact:
    name: M360 Team
  license:
    name: Proprietary
servers:
  - url: /api/v1
    description: API v1 base path
security:
  - BearerAuth: []
...
paths:
...
  /facilities:
    get:
      tags: [Facilities]
      summary: List facilities
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, closed, defaulted, restructured]
        - name: organization_id
          in: query
          schema:
            type: string
            format: uuid
        - name: delinquency
          in: query
          schema:
            type: string
            enum: [current, par_30, par_60, par_90, par_180, write_off]
      responses:
        '200':
          description: Paginated list of facilities
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacilityListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Facilities]
      summary: Create a facility
      description: Requires admin or manager role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacilityCreateRequest'
      responses:
        '201':
          description: Facility created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Facility'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /facilities/{id}:
    get:
      tags: [Facilities]
      summary: Get facility by ID
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Facility details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Facility'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /facilities/{id}/schedule:
    get:
      tags: [Facilities]
      summary: Get facility repayment schedule
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Repayment schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  facility_id:
                    type: string
                    format: uuid
                  schedule:
                    type: array
                    items:
                      $ref: '#/components/schemas/RepaymentScheduleItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /facilities/{id}/payments:
    post:
      tags: [Facilities]
      summary: Record a payment on a facility
      parameters:
        - $ref: '#/components/parameters/IDParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordPaymentRequest'
      responses:
        '200':
          description: Payment recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Payment recorded successfully
                  facility:
                    $ref: '#/components/schemas/Facility'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Collections
  # ──────────────────────────────────────────────
...
  /documents:
    post:
      tags: [Documents]
      summary: Upload a document
      description: |
        Upload a file (max 10MB). Allowed types: PDF, JPEG, PNG, DOC, DOCX, XLS, XLSX.
        Uses multipart/form-data with fields: file, entity_type, entity_id, name, category (optional).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, entity_type, entity_id, name]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload (max 10MB)
                entity_type:
                  type: string
                  description: Type of entity the document belongs to (e.g. application, organization)
                entity_id:
                  type: string
                  format: uuid
                  description: ID of the entity
                name:
                  type: string
                  description: Document name
                category:
                  type: string
                  description: Document category (optional)
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Get document by ID
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Documents]
      summary: Delete a document
      description: Requires admin or manager role.
      parameters:
        - $ref: '#/components/parameters/IDParam'
      responses:
        '200':
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /documents/entity/{type}/{id}:
    get:
      tags: [Documents]
      summary: List documents by entity
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
          description: Entity type (e.g. application, organization)
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Entity ID
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

  # ──────────────────────────────────────────────
  # Activities
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/refresh
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    LimitParam:
      name: limit
      in: query
      description: Number of items per page (default 20)
      schema:
        type: integer
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
    MessageResponse:
      type: object
      properties:
        message:
          type: string

    # ── Auth ──
...
    Facility:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reference_number:
          type: string
        application_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        principal_amount:
          type: number
          format: double
        profit_amount:
          type: number
          format: double
        total_amount:
          type: number
          format: double
        outstanding_balance:
          type: number
          format: double
        profit_rate:
          type: number
          format: double
        tenor_months:
          type: integer
        payment_frequency:
          type: string
          description: "monthly, quarterly, etc."
        disbursement_date:
          type: string
          format: date-time
        maturity_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, closed, defaulted, restructured]
        delinquency:
          type: string
          enum: [current, par_30, par_60, par_90, par_180, write_off]
        assigned_officer_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FacilityCreateRequest:
      type: object
      required: [application_id, organization_id, product_id, principal_amount, profit_amount, tenor_months, payment_frequency, disbursement_date]
      properties:
        application_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        principal_amount:
          type: number
          format: double
        profit_amount:
          type: number
          format: double
        tenor_months:
          type: integer
        payment_frequency:
          type: string
        disbursement_date:
          type: string
          format: date-time
    FacilityListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Facility'
        pagination:
          type: object
          properties:
            total:
              type: integer
            limit:
              type: integer
            offset:
              type: integer
    RepaymentScheduleItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        facility_id:
          type: string
          format: uuid
        installment_number:
          type: integer
        due_date:
          type: string
          format: date-time
        principal_amount:
          type: number
          format: double
        profit_amount:
          type: number
          format: double
        total_amount:
          type: number
          format: double
        paid_amount:
          type: number
          format: double
        paid_date:
          type: string
          format: date-time
          nullable: true
        is_paid:
          type: boolean
        is_overdue:
          type: boolean
        created_at:
          type: string
          format: date-time
    RecordPaymentRequest:
      type: object
      required: [installment_number, paid_amount, payment_date]
      properties:
        installment_number:
          type: integer
        paid_amount:
          type: number
          format: double
        payment_date:
          type: string
          format: date-time

    # ── Collection ──
...
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entity_type:
          type: string
        entity_id:
          type: string
          format: uuid
        name:
          type: string
        file_path:
          type: string
        file_size:
          type: integer
          format: int64
        mime_type:
          type: string
        category:
          type: string
          nullable: true
        uploaded_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time

    # ── Activity ──
...